requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.2
pydantic==2.7.4
python-dotenv==1.0.1
PyYAML==6.0.2
sqlalchemy==2.0.23
trafilatura==1.8.0
readability-lxml==0.8.4.1
beautifulsoup4==4.12.3
lxml==5.2.1
sentence-transformers==2.7.0
scikit-learn==1.5.2
numpy==1.26.4
tenacity==8.5.0
openai==1.41.0
orjson==3.10.4
jinja2==3.1.4
